use mixin for schema-field initialization and update, use one-way-input for text
