+++ package.json
@@ -37,6 +37,7 @@
     "ember-disable-prototype-extensions": "^1.0.0",
     "ember-disable-proxy-controllers": "^1.0.0",
     "ember-export-application-global": "^1.0.4",
+    "ember-one-way-input": "0.1.1",
     "ember-radio-button": "1.0.7",
     "ember-resolver": "^2.0.2",
     "ember-suave": "^1.2.0",
